requirements.txt
Django~=3.2.5
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.5.0
django-environ
psycopg2-binary==2.8.6
PyJWT~=2.4.0
python-dateutil~=2.8.2